+++ src/clj/rems/context.clj
@@ -18,3 +18,5 @@
 (def ^:dynamic ^{:doc "Tempura object initialized with user's preferred language."} *tempura*)
 
 (def ^:dynamic ^{:doc "User's preferred language."} *lang*)
+
+(def ^:dynamic ^{:doc "Ongoing HTTP request if any."} *request*)